+++ src/typing_extensions.py
@@ -644,14 +644,17 @@ else:
         def __subclasscheck__(cls, other):
             if cls is Protocol:
                 return type.__subclasscheck__(cls, other)
-            if not isinstance(other, type):
-                # Same error message as for issubclass(1, int).
-                raise TypeError('issubclass() arg 1 must be a class')
             if (
                 getattr(cls, '_is_protocol', False)
                 and not _allow_reckless_class_checks()
             ):
-                if not cls.__callable_proto_members_only__:
+                if not isinstance(other, type):
+                    # Same error message as for issubclass(1, int).
+                    raise TypeError('issubclass() arg 1 must be a class')
+                if (
+                    not cls.__callable_proto_members_only__
+                    and cls.__dict__.get("__subclasshook__") is _proto_hook
+                ):
                     raise TypeError(
                         "Protocols with non-method members don't support issubclass()"
                     )
@@ -752,12 +755,8 @@ else:
                 if '__subclasshook__' not in cls.__dict__:
                     cls.__subclasshook__ = _proto_hook
 
-                # We have nothing more to do for non-protocols...
-                if not cls._is_protocol:
-                    return
-
-                # ... otherwise prohibit instantiation.
-                if cls.__init__ is Protocol.__init__:
+                # Prohibit instantiation for protocol classes
+                if cls._is_protocol and cls.__init__ is Protocol.__init__:
                     cls.__init__ = _no_init
 
     else:
@@ -847,12 +846,8 @@ else:
                 if '__subclasshook__' not in cls.__dict__:
                     cls.__subclasshook__ = _proto_hook
 
-                # We have nothing more to do for non-protocols.
-                if not cls._is_protocol:
-                    return
-
-                # Prohibit instantiation
-                if cls.__init__ is Protocol.__init__:
+                # Prohibit instantiation for protocol classes
+                if cls._is_protocol and cls.__init__ is Protocol.__init__:
                     cls.__init__ = _no_init